@@ -10,7 +10,7 @@ description: A flutter boilerplate project created using MobX and Provider.
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.6 <4.0.0'
+  sdk: '>=3.3.19 <4.0.0'
   
 dependencies:
   flutter:
@@ -29,20 +29,20 @@ dependencies:
   # E
   event_bus: ^2.0.0
   # F
-  flutter_mobx: ^2.0.6+5
+  flutter_mobx: ^2.2.1+1
   # G
   get_it: ^7.2.0
-  google_fonts: ^4.0.3
+  google_fonts: ^6.2.1
   # H
-  http: ^0.13.5
+  http: ^1.2.1
   # I
-  intl: ^0.18.0
+  intl: ^0.19.0
   # J
   json_serializable: ^6.6.1
   # K
   # L
   # M
-  material_dialog: ^1.0.0-nullsafety.1
+  material_dialogs: ^1.1.4
   mobx: ^2.1.4
   # N
   # O
@@ -75,7 +75,7 @@ dev_dependencies:
   flutter_launcher_icons: ^0.13.1
   mobx_codegen: ^2.2.0
   build_runner: ^2.3.3
-  analyzer: ^5.10.0
+  analyzer: ^6.4.1
 
 flutter_icons:
   image_path: "assets/icons/ic_launcher.png"
