@@ -54,7 +54,7 @@ class AppThemeData {
           _lightFillColor.withOpacity(0.80),
           _darkFillColor,
         ),
-        contentTextStyle: _textTheme.subtitle1!.apply(color: _darkFillColor),
+        contentTextStyle: _textTheme.titleLarge!.apply(color: _darkFillColor),
       ),
     );
   }
@@ -98,15 +98,15 @@ class AppThemeData {
   static const _bold = FontWeight.w700;
 
   static final TextTheme _textTheme = TextTheme(
-    headline4: GoogleFonts.montserrat(fontWeight: _bold, fontSize: 20.0),
-    caption: GoogleFonts.oswald(fontWeight: _semiBold, fontSize: 16.0),
-    headline5: GoogleFonts.oswald(fontWeight: _medium, fontSize: 16.0),
-    subtitle1: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 16.0),
-    overline: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 12.0),
-    bodyText1: GoogleFonts.montserrat(fontWeight: _regular, fontSize: 14.0),
-    subtitle2: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 14.0),
-    bodyText2: GoogleFonts.montserrat(fontWeight: _regular, fontSize: 16.0),
-    headline6: GoogleFonts.montserrat(fontWeight: _bold, fontSize: 16.0),
-    button: GoogleFonts.montserrat(fontWeight: _semiBold, fontSize: 14.0),
+    headlineLarge: GoogleFonts.montserrat(fontWeight: _bold, fontSize: 20.0),
+    bodyLarge: GoogleFonts.oswald(fontWeight: _semiBold, fontSize: 16.0),
+    headlineMedium: GoogleFonts.oswald(fontWeight: _medium, fontSize: 16.0),
+    titleLarge: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 16.0),
+    labelMedium: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 12.0),
+    bodyMedium: GoogleFonts.montserrat(fontWeight: _regular, fontSize: 14.0),
+    titleMedium: GoogleFonts.montserrat(fontWeight: _medium, fontSize: 14.0),
+    bodySmall: GoogleFonts.montserrat(fontWeight: _regular, fontSize: 16.0),
+    headlineSmall: GoogleFonts.montserrat(fontWeight: _bold, fontSize: 16.0),
+    labelLarge: GoogleFonts.montserrat(fontWeight: _semiBold, fontSize: 14.0),
   );
 }
