@@ -77,7 +77,7 @@ class _PostListScreenState extends State<PostListScreen> {
         maxLines: 1,
         overflow: TextOverflow.ellipsis,
         softWrap: false,
-        style: Theme.of(context).textTheme.subtitle1,
+        style: Theme.of(context).textTheme.titleLarge,
       ),
       subtitle: Text(
         '${_postStore.postList?.posts?[position].body}',
