@@ -7,7 +7,8 @@ import 'package:boilerplate/utils/locale/app_localization.dart';
 import 'package:boilerplate/utils/routes/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:material_dialog/material_dialog.dart';
+import 'package:material_dialogs/material_dialogs.dart';
+import 'package:material_dialogs/shared/types.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -85,59 +86,47 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 
   _buildLanguageDialog() {
-    _showDialog<String>(
+    Dialogs.materialDialog(
+      color: Theme.of(context).scaffoldBackgroundColor,
       context: context,
-      child: MaterialDialog(
-        borderRadius: 5.0,
-        enableFullWidth: true,
-        title: Text(
-          AppLocalizations.of(context).translate('home_tv_choose_language'),
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 16.0,
-          ),
-        ),
-        headerColor: Theme.of(context).primaryColor,
-        backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-        closeButtonColor: Colors.white,
-        enableCloseButton: true,
-        enableBackButton: false,
-        onCloseButtonClicked: () {
-          Navigator.of(context).pop();
-        },
-        children: _languageStore.supportedLanguages
-            .map(
-              (object) => ListTile(
-                dense: true,
-                contentPadding: EdgeInsets.all(0.0),
-                title: Text(
-                  object.language!,
-                  style: TextStyle(
-                    color: _languageStore.locale == object.locale
-                        ? Theme.of(context).primaryColor
-                        : _themeStore.darkMode
-                            ? Colors.white
-                            : Colors.black,
-                  ),
-                ),
-                onTap: () {
-                  Navigator.of(context).pop();
-                  // change user language based on selected locale
-                  _languageStore.changeLanguage(object.locale!);
-                },
+      customViewPosition: CustomViewPosition.BEFORE_ACTION,
+      customView: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16.0),
+        child: Column(
+          children: [
+            Text(
+              AppLocalizations.of(context).translate('home_tv_choose_language'),
+              style: TextStyle(
+                color: _themeStore.darkMode ? Colors.white : Colors.black,
+                fontSize: 16.0,
               ),
-            )
-            .toList(),
+            ),
+            ..._languageStore.supportedLanguages
+                .map(
+                  (object) => ListTile(
+                    dense: true,
+                    contentPadding: EdgeInsets.all(0.0),
+                    title: Text(
+                      object.language!,
+                      style: TextStyle(
+                        color: _languageStore.locale == object.locale
+                            ? Theme.of(context).primaryColor
+                            : _themeStore.darkMode
+                                ? Colors.white
+                                : Colors.black,
+                      ),
+                    ),
+                    onTap: () {
+                      Navigator.of(context).pop();
+                      // change user language based on selected locale
+                      _languageStore.changeLanguage(object.locale!);
+                    },
+                  ),
+                )
+                .toList()
+          ],
+        ),
       ),
     );
-  }
-
-  _showDialog<T>({required BuildContext context, required Widget child}) {
-    showDialog<T>(
-      context: context,
-      builder: (BuildContext context) => child,
-    ).then<void>((T? value) {
-      // The value passed to Navigator.pop() or null.
-    });
   }
 }
